Rename MORR.Core.Data.Capture.Video.WinAPI to MORR.Core.Data.Capture.Video.Desktop, remove unused IVideoCapture
