+++ application/Core/MORR/Data/Capture/Video/Desktop/DesktopCaptureConfiguration.cs
@@ -3,7 +3,7 @@ using System.Text.Json;
 using MORR.Core.Configuration;
 using MORR.Shared.Configuration;
 
-namespace MORR.Core.Data.Capture.Video.WinAPI
+namespace MORR.Core.Data.Capture.Video.Desktop
 {
     public class DesktopCaptureConfiguration : IConfiguration
     {
